refactor(get_context): split pipeline.rs into second_hop and task_signals
Move second-hop graph expansion helpers (should_expand_second_hop,
select_second_hop_seeds, merge_expansions) into a new second_hop module.
Move failing-test linkage helpers (hydrate_failing_test_links,
test_name_matches_signal) into task_signals next to TaskSignals.

Pure refactor — no behavior change. Cuts pipeline.rs from 848 to 747
lines. Updates the one external test importer to reach
hydrate_failing_test_links through its new home.
